+++ CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Delete `gen.go_options.no_default_modifiers` setting.
 - Delete `lint.group` setting.
 - Delete `harbormaster` global flag.
-- Rename `create.dir_to_base_package` -> `create.dir_to_package`.
+- Refactor `create.dir_to_base_package` to the list `create.packages` See
+  the documentation for more details..
 
 
 ## [0.5.0] - 2018-07-26